+++ lib/chemicalml/cml/reaction_list.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require "lutaml/model"
+
+module ChemicalML
+  module Cml
+    # CML `<reactionList>` element. Container for `<reaction>` children.
+    class ReactionList < Lutaml::Model::Serializable
+      attribute :reactions, Reaction, collection: true
+
+      xml do
+        root "reactionList"
+        map_element "reaction", to: :reactions
+      end
+    end
+  end
+end