+++ lib/chemicalml/errors.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ChemicalML
+  class Error < StandardError; end
+
+  class ParseError < Error; end
+end